requirements.txt
web3
requests
colorama
prettytable
python-dotenv
